gitlab CI: drop MESON_SKIP_TEST variables
In the no-check:meson job, the ninja arg was "dist" so the test would be
run as part of that anyway (and skipped, since we didn't have check).

In the no-doxygen-check:meson job, the ninja arg was zero so the test
would be skipped but since we don't have check we might as well just
run it as empty test suite.

And the same applies to the scan-build job, running the test shouldn't
hurt here.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>