Include all include files in dist tarballs
Otherwise, the distribution tarballs will be generated based on what's
in libevdev_la_SOURCES, which only includes the headers for the
operating system the maintainer happens to be running "make dist"
from.  As a result of this, e.g. the 1.13.0 tarball only includes the
headers for Linux, making the tarball unbuildable for FreeBSD.

"meson dist" was already doing the right thing.

Signed-off-by: Alyssa Ross <hi@alyssa.is>